portal/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
PyJWT==2.8.0
cryptography==41.0.3
bcrypt==4.0.1
email-validator==2.0.0
Werkzeug==2.3.7
psutil==5.9.5
requests==2.31.0
urllib3==2.0.4 